Fixed weird issue with music volume (notes)
For some reason when starting up the game, the music/audiosource volume would be low, but upon switching scenes/rooms the audiosource volume would increase to full as it should. For some reason using Time.unscaledDeltaTime in FadeTrack for the MusicManager on game startup was messing with startup volume, so I created a duplicate that uses Time.deltaTime as usual and that seems to have fixed it. No clue why unscaledDeltaTime would cause an issue on startup, unless it's an editor-only thing?