Create global enemy pool (notes)
Removed manual list of possible enemies from every Room prefab. Now each enemy instance is set to be a child of the Room that instanced it, then each spawned enemy looks for the RoomBehavior script in its parent. Then it automatically assigns the variable and is able to remove itself from the RoomBehavior's Spawned Enemies list so that the player can clear the room (by emptying the list to 0)