Create global pickups pool (notes)
Did the same thing for pickups, so now all rooms can reference a pickup from a global list instead of having a dedicated child in each prefab. Like with enemies, this is possible partly due to having a GlobalOnDestroySounds script, and eliminating the old local OnDestroySounds that was part of each room prefab. Nice