Some weird player projectile bug (notes)
For some reason after going through some rooms the player keeps shooting "ghost" projectiles that aren't showing up as active in the Projectile Holder list, and they only last for a fraction of the time of a normal projectile. There was also an issue with some projectiles not resetting scale or velocity so they'd shoot really fast or end too quickly.