Projectile bug fix (notes)
I added the playerEnteredNewRoom action to player projectiles so that they'd all get disabled upon player entering new room, but the desable method also included adding the projectiles to the Player's projectile list. I didn't check to see if projectiles were active or not, so it was adding a bunch of inactive duplicate projectiles to the Player's list, which had weird interactions and wouldn't fire correctly when 2 of the "same" projectiles were fired consecutively. Adding a simple if statement to check if the projectile is activeInHierarchy before disabling it using the playerENteredNewRoom action fixed the issue.